requirements.txt
fastapi
uvicorn
python-multipart
requests
PyPDF2
python-docx
sentence-transformers
faiss-cpu
numpy
pydantic
aiofiles
beautifulsoup4
lxml
tiktoken
transformers
torch
mistralai
python-dotenv
pandas
openpyxl
pytesseract
Pillow
python-pptx
oletools
PyMuPDF
langchain-text-splitters
httpx
demjson